update avatar logic
