+++ tools/tools.py
@@ -4,11 +4,11 @@ import time
 import json
 import pickle
 
-def download(file_path, url):
+def download(file_path, url, timeout=120):
 	headers = {
 		"User-Agent": "Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 			(KHTML, like Gecko) Chrome/63.0.3239.132 Safari/537.36 QIHU 360SE",
 		}
-	r = requests.get(url, headers=headers)
+	r = requests.get(url, headers=headers, timeout=timeout)
 	with open(file_path, 'wb') as f:
 		f.write(r.content)
 
@@ -40,15 +40,31 @@ def imageWebsite(href:str):
 	host = url.split("/")[0]
 	return host
 
-def script2caption(script:str):
-	sections = script.split("\n")
-	if len(sections) > 3:
-		captions = sections
-	else:
-		captions = []
-		for section in sections:
-			captions += section.split("。")
+def getTextArrayFromArray(textArrays: [str], sep:str='。'):
+	newArray = []
+	for text in textArrays:
+		subArray = text.split(sep)
+		if len(subArray) > 1:
+			newArray += [v + sep for v in subArray[:-1]]
+			newArray.append(subArray[-1])
+		else:
+			newArray.append(text)
+	return newArray
 
+
+def script2caption(script:str):
+	captions = getTextArrayFromArray(script.split("\n"), "。")
+	captions = getTextArrayFromArray(captions, "？")
+	captions = getTextArrayFromArray(captions, "！")
 	captions = filter(lambda s: len(s) > 0, captions)
 	return captions
+
+def tryHandle(func, max_try:int = 3, **args):
+	try:
+		return func(**args)
+	except Exception as e:
+		if max_try > 0:
+			return tryHandle(func, max_try=max_try -1, **args )
+		else:
+			raise e
